backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
email-validator
python-dotenv